ShazamAPI-main/requirements.txt
Flask
flask-cors
werkzeug
mutagen
python-dotenv
boto3
botocore
uuid
schedule
pydub
requests